Combine mutations and listening
They serve the same purpose.
